[fuchsia] Updates the FIDL bindings in Fuchsia-specific code
Fuchsia has a code snippet that uses its FIDL IPC mechanism
to request time zone information from the host.

The generated FIDL bindings have been changed since this code
was first written from the "high level" (HL) CPP bindings to the
"new" CPP bindings.

We are now migrating external code to remove the HL CPP bindings
to allow FIDL to continue development.  No other platforms should
be affected, since the changes are in the sections of the code
that is conditionally compiled only when Fuchsia is the target
platform.